requirements.txt
# Datenhandling
pandas  # main package for data handling
xlrd  # for reading excel files
pyyaml  # for reading yaml files
# numpy
# pdfplumber
# openpyxl


# Web scraping
selenium
# undetected_chromedriver
# webdriver-manager

# Gui
# pyqt5  # old package
# pyqt6  # for GUI
# pyqt6-tools  # for GUI design

# Image processing - OCR
# opencv-python
# easyocr
# matplotlib

#--extra-index-url https://pypi.ngc.nvidia.com
#nvidia-pyindex
#nvidia-cuda-runtime-cu12

# pytesseract

# Standard packages
# pip~=24.3.1
# wheel~=0.45.1
# setuptools~=75.7.0
